# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "ProjectController_getAll",
        "summary": "",
        "description": "Get all project",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectResponseDto" } }
              }
            }
          }
        },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "ProjectController_create",
        "summary": "",
        "description": "Create new project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/CreateProjectRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectResponseDto" } } }
          },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectController_findById",
        "summary": "",
        "description": "Find by ID project",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectResponseDto" } } }
          },
          "404": { "description": "There isn't any project with id" }
        },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "ProjectController_update",
        "summary": "",
        "description": "Update project",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/UpdateProjectRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectResponseDto" } } }
          },
          "404": { "description": "There isn't any project with id" },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "ProjectController_delete",
        "summary": "",
        "description": "Delete project",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" }, "404": { "description": "There isn't any project with id" } },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/search": {
      "get": {
        "operationId": "ProjectController_search",
        "summary": "",
        "description": "Search Project",
        "parameters": [{ "name": "Project name", "required": true, "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectResponseDto" } }
              }
            }
          }
        },
        "tags": ["Projects"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/loggedDays": {
      "get": {
        "operationId": "LoggedDayController_find",
        "summary": "",
        "description": "Find logged days",
        "parameters": [
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Choose \"Date from\" (rrrr-mm-dd)",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Choose \"Date to\" (rrrr-mm-dd)",
            "schema": { "format": "date-time", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/GetLoggedDaysFilterRequestDto" } }
              }
            }
          }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "LoggedDayController_create",
        "summary": "",
        "description": "Create new logged day",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateLoggedDayRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CreateLoggedDayResponseDto" } }
            }
          }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      }
    },
    "/loggedDays/{id}": {
      "get": {
        "operationId": "LoggedDayController_findById",
        "summary": "",
        "description": "Find by ID logged day",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CreateLoggedDayResponseDto" } }
            }
          },
          "500": { "description": "Internal server error" }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "LoggedDayController_update",
        "summary": "",
        "description": "Update logged day",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateLoggedDayRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UpdateLoggedDayRequestDto" } }
            }
          },
          "404": { "description": "There isn't any logged day with id" },
          "500": { "description": "Internal server error" }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "LoggedDayController_delete",
        "summary": "",
        "description": "Delete logged day",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "" },
          "404": { "description": "There isn't any logged day with this id" }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      }
    },
    "/loggedDays/byIds": {
      "delete": {
        "operationId": "LoggedDayController_deleteByIds",
        "summary": "",
        "description": "Delete logged days",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteByIdsRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "Undeleted ids",
            "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } }
          }
        },
        "tags": ["loggedDays"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/statistic": {
      "get": {
        "operationId": "StatisticController_getByGrouped",
        "summary": "",
        "description": "Get statistic",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "UserId",
            "schema": { "type": "string" }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "ProjectId",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/StatisticDiagramResponseDto" } }
              }
            }
          }
        },
        "tags": ["Statistics"],
        "security": [{ "bearer": [] }]
      }
    },
    "/statistic/count": {
      "get": {
        "operationId": "StatisticController_getCount",
        "summary": "",
        "description": "Get count statistic",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "UserId",
            "schema": { "type": "string" }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "ProjectId",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StatisticHeaderResponseDto" } }
            }
          }
        },
        "tags": ["Statistics"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": { "title": "BACKEND", "description": "REST API", "version": "1.0.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "DepartmentResponseDto": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
        "required": ["id", "name"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string" },
          "skills": { "type": "string" },
          "lastSessionDateTime": { "type": "string" },
          "avatarPath": { "type": "string" },
          "phone": { "type": "string" },
          "department": { "$ref": "#/components/schemas/DepartmentResponseDto" }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt",
          "firstName",
          "lastName",
          "role",
          "skills",
          "lastSessionDateTime",
          "avatarPath",
          "phone",
          "department"
        ]
      },
...
      "DeleteByIdsRequestDto": {
        "type": "object",
        "properties": {
          "ids": {
            "example": ["9322c384-fd8e-4a13-80cd-1cbd1ef95ba8", "986dcaf4-c1ea-4218-b6b4-e4fd95a3c28e"],
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["ids"]
      },
...
      "ClientResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "website": { "type": "string" },
          "avatarPath": { "type": "string" },
          "createdAt": { "type": "string" }
        },
        "required": ["id", "name", "website", "avatarPath", "createdAt"]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "createdAt": { "type": "string" },
          "imagePath": { "type": "string" },
          "client": { "$ref": "#/components/schemas/ClientResponseDto" },
          "users": { "type": "array", "items": { "$ref": "#/components/schemas/UserResponseDto" } }
        },
        "required": ["id", "name", "createdAt", "imagePath", "client", "users"]
      },
      "CreateProjectRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Project Name", "example": "Esta" },
          "clientId": {
            "type": "string",
            "description": "Client ID",
            "example": "9878f24c-6b96-43cd-a1c4-69c66c875827"
          },
          "file": { "type": "string", "format": "binary" },
          "performerIds": {
            "example": ["9878f24c-6b96-43cd-a1c4-69c66c875827"],
            "description": "Array of performer ids",
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["name", "clientId"]
      },
      "UpdateProjectRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Project Name", "example": "Esta" },
          "clientId": {
            "type": "string",
            "description": "Client ID",
            "example": "9878f24c-6b96-43cd-a1c4-69c66c875827"
          },
          "file": { "type": "string", "format": "binary" },
          "performerIds": {
            "example": ["9878f24c-6b96-43cd-a1c4-69c66c875827"],
            "description": "Array of performer ids",
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["name", "clientId"]
      },
...
      "GetLoggedDaysFilterRequestDto": {
        "type": "object",
        "properties": {
          "dateFrom": { "format": "date-time", "type": "string", "description": "Choose \"Date from\" (rrrr-mm-dd)" },
          "dateTo": { "format": "date-time", "type": "string", "description": "Choose \"Date to\" (rrrr-mm-dd)" }
        }
      },
      "CreateLoggedDayResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "hours": { "type": "number" },
          "minutes": { "type": "number" },
          "description": { "type": "string" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "date": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserResponseDto" },
          "project": { "$ref": "#/components/schemas/ProjectResponseDto" },
          "department": { "$ref": "#/components/schemas/DepartmentResponseDto" },
          "job": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        },
        "required": [
          "id",
          "hours",
          "minutes",
          "description",
          "userId",
          "projectId",
          "date",
          "user",
          "project",
          "department",
          "job",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateLoggedDayRequestDto": {
        "type": "object",
        "properties": {
          "hours": { "type": "number", "description": "hours worked", "example": 2 },
          "minutes": { "type": "number", "description": "minutes worked", "example": 30 },
          "description": { "type": "string", "description": "task description", "example": "super hard task" },
          "projectId": {
            "type": "string",
            "description": "project uuid",
            "example": "cd9d51b5-af60-4736-84ac-f4c6784d69a3"
          },
          "jobId": { "type": "string", "description": "job uuid", "example": "cd9d51b5-af60-4736-84ac-f4c6784d69a3" },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "day timestamp",
            "example": "2022-10-06T06:24:38.584Z"
          }
        },
        "required": ["hours", "minutes", "description", "projectId", "jobId", "date"]
      },
      "UpdateLoggedDayRequestDto": {
        "type": "object",
        "properties": {
          "hours": { "type": "number", "description": "hours worked", "example": 2 },
          "minutes": { "type": "number", "description": "minutes worked", "example": 30 },
          "description": { "type": "string", "description": "task description", "example": "super hard task" },
          "projectId": {
            "type": "string",
            "description": "project uuid",
            "example": "cd9d51b5-af60-4736-84ac-f4c6784d69a3"
          },
          "jobId": { "type": "string", "description": "job uuid", "example": "cd9d51b5-af60-4736-84ac-f4c6784d69a3" },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "day timestamp",
            "example": "2022-10-06T06:24:38.587Z"
          }
        },
        "required": ["hours", "minutes", "description", "projectId", "jobId", "date"]
      },
      "StatisticDiagramResponseDto": {
        "type": "object",
        "properties": {
          "date": { "type": "string" },
          "hours": { "type": "number" },
          "minutes": { "type": "number" },
          "name": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "path": { "type": "string" }
        },
        "required": ["date", "hours", "minutes", "name", "firstName", "lastName", "path"]
      },
      "StatisticHeaderResponseDto": {
        "type": "object",
        "properties": {
          "projectsCount": { "type": "number", "description": "Projects Count", "example": 4 },
          "usersCount": { "type": "number", "description": "Users Count", "example": 5 },
          "hours": { "type": "number", "description": "All hours", "example": 124 }
        },
        "required": ["projectsCount", "usersCount", "hours"]
      },
...
    }
  }
}